+++ inherited-rules/implicitly-included/node-recommended-module.js
@@ -1,3 +1,4 @@
+"use strict";
 module.exports = {
   rules: {
     "node/no-deprecated-api": "error",