+++ inherited-rules/implicitly-included/sonarjs-recommended.js
@@ -1,3 +1,4 @@
+"use strict";
 module.exports = {
   rules: {
     "sonarjs/no-all-duplicated-branches": "error",